# backend/docs/swagger.yaml
basePath: /api
definitions:
...
  gymtrack-backend_internal_domain_models.UserProfile:
    properties:
      age:
        type: integer
      certifications:
        description: Trainer specific fields
        type: string
      clientList:
        description: List of athlete IDs
        items:
          type: string
        type: array
      fitnessGoals:
        type: string
      height:
        type: integer
      name:
        type: string
      specializations:
        type: string
      trainerAssignment:
        description: Athlete's trainer ID
        type: string
      weight:
        type: integer
    type: object
  gymtrack-backend_internal_domain_models.UserRole:
    enum:
    - trainer
    - athlete
    type: string
    x-enum-varnames:
    - RoleTrainer
    - RoleAthlete
...
  internal_api_handlers.UpdateProfileRequest:
    description: UpdateProfileRequest is used to modify user profile information.
    properties:
      profile:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.UserProfile'
        description: |-
          Profile contains the updated profile information.
          @Description The user's profile data including personal details and role-specific fields.
    required:
    - profile
    type: object
...
  internal_api_handlers.UserResponse:
    description: UserResponse contains all user information except the password hash.
    properties:
      createdAt:
        description: |-
          CreatedAt is the timestamp when the user account was created.
          @Description The creation timestamp of the user account.
          @Example 2024-01-01T00:00:00Z
        type: string
      email:
        description: |-
          Email is the user's email address.
          @Description The user's registered email address.
          @Example user@example.com
        type: string
      profile:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.UserProfile'
        description: |-
          Profile contains user-specific profile information.
          @Description The user's profile data including personal details and role-specific fields.
      role:
        allOf:
        - $ref: '#/definitions/gymtrack-backend_internal_domain_models.UserRole'
        description: |-
          Role is the user's role in the system (trainer or athlete).
          @Description The user's role determining their permissions and capabilities.
          @Example athlete
          @Enum trainer,athlete
      updatedAt:
        description: |-
          UpdatedAt is the timestamp when the user was last updated.
          @Description The last update timestamp of the user account.
          @Example 2024-01-01T00:00:00Z
        type: string
      userId:
        description: |-
          UserID is the unique identifier for the user.
          @Description The unique UUID of the user.
          @Example 123e4567-e89b-12d3-a456-426614174000
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Fitness tracking API for personal trainers and athletes
  title: GymTrack API
  version: "1.0"
paths:
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Returns the profile information of the user associated with the
        provided JWT token.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user profile
          schema:
            $ref: '#/definitions/internal_api_handlers.UserResponse'
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Updates the profile information for the authenticated user. Returns
        the updated user data.
      parameters:
      - description: Update profile request containing the updated profile data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_api_handlers.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user profile
          schema:
            $ref: '#/definitions/internal_api_handlers.UserResponse'
        "400":
          description: Bad request - Invalid input data
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized - Invalid or missing JWT token
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update current user profile
      tags:
      - User
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"